Update image used for live cells
I was previously using an image of a Minecraft "creep" that I had
snagged off google. I never really intended on keeping it for any
duration of time, but somehow it managed to find itself used as even the
icon for the early iterations of this project. I am finally going back
and enhancing the graphics in an attempt prepare the project for my
portfolio.